Replace uses of IDbContext.CreateCommand
with DbConnection.CreateCommand

Only two members of IDbContext are in use: CreateCommand and CreateConnection.

There is only one call site that uses IDbContext.CreateCommand and this can be
trivially replaced with IDbConnection.CreateCommand. This change satisfies the
compiler but breaks several unit tests, which now require
DbConnection.CreateCommand to be virtual.